+++ include/em_persistent_state.h
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef _EM_PERSISTENT_STATE_H_
+#define _EM_PERSISTENT_STATE_H_
 
 #include <stdint.h>
 #include <stdlib.h>
@@ -8,6 +9,7 @@
 
 #include "em_log.h"
 #include "em_defs.h"
+#include "em_optional.h"
 #include "em_list.h"
 #include "em_sync_value.h"
 
@@ -20,7 +22,8 @@ class EmPersistentId;
 class EmPersistentState;
 class EmPersistentValueBase;
 class EmPersistentValueIterator;
-bool _itemsMatch(const EmPersistentValueBase& pv1, 
+typedef EmIterator<EmPersistentValueBase> EmPSIterator;
+bool itemsMatch_(const EmPersistentValueBase& pv1, 
                  const EmPersistentValueBase& pv2);
 
 /***
@@ -28,7 +31,7 @@ bool _itemsMatch(const EmPersistentValueBase& pv1,
 ***/
 class EmPersistentValueList: public EmList<EmPersistentValueBase> {
 public:
-    EmPersistentValueList() : EmList<EmPersistentValueBase>(_itemsMatch) {}
+    EmPersistentValueList() : EmList<EmPersistentValueBase>(itemsMatch_) {}
 };
 
 /***
@@ -44,20 +47,20 @@ public:
         void setup() {
             // First way to initialize Persistent State
             EmPersistentValueList values;
-            values.Append(floatVal);
-            values.Append(intVal);
-            values.Append(textVal);
-            PS.Init(values, false);
+            values.append(floatVal);
+            values.append(intVal);
+            values.append(textVal);
+            PS.begin(values, false);
         }
 
         void setup_() {
             // Second way to initialize Persistent State
             // NOTE: this way we cannot delete old unused values but we 
             //       spare memory by not filling an 'EmPersistentValueList' 
-            if (PS.Init() >= 0) {
-                PS.Add(floatVal);
-                PS.Add(intVal);
-                PS.Add(textVal);
+            if (PS.begin() >= 0) {
+                PS.add(floatVal);
+                PS.add(intVal);
+                PS.add(textVal);
             }
         }
 
@@ -72,16 +75,22 @@ public:
         }
 ***/
 class EmPersistentState: public EmLog {
-    friend class EmPersistentValueBase;
     friend class EmPersistentId;
+    friend class EmPersistentValueBase;
+    friend class EmPersistentValueIterator;
 public:    
     const static EmPersistentId c_HeaderId; 
     const static EmPersistentId c_FooterId;
     const static int c_MinSize = 12;
 
-    EmPersistentState(EmLogLevel logLevel = EmLogLevel::none,
-                      ps_address_t beginIndex = EEPROM.begin(),
-                      ps_address_t endIndex = EEPROM.end());
+#ifdef EM_EEPROM
+    EmPersistentState(EmOptional<ps_address_t> beginIndex = emUndefined,
+                      EmOptional<ps_address_t> endIndex = emUndefined,
+                      EmLogLevel logLevel = EmLogLevel::global);
+#else
+    EmPersistentState(ps_size_t size = 1024,
+                      EmLogLevel logLevel = EmLogLevel::global);
+#endif
 
     // NOTE: keep destructor and class without virtual functions to avoid extra RAM consumption
     ~EmPersistentState() {
@@ -91,103 +100,127 @@ public:
     //
     // Return the persistent state stored values count or -1 if persistent state 
     // has not been successfully initialized.
-    int Init();
+    int begin();
 
-    // Initialize the persistent state by loading 'values' list, 
+    // Initialize the persistent state by loading 'values' iterator, 
     // setting current values to the requested 'values'.
     //
     // Return the persistent state stored values count or -1 if persistent state 
     // has not been successfully initialized.
     //
     // If 'removeUnusedValues' is set to true the storage will remove 
-    // values that are not in the desired 'values' list.
+    // values that are not in the desired 'values' iterator.
     //
     // NOTE:
     //   appending values to the 'values' list after the 'Init' call has no effect.
     //   Use the persistent state 'Add' method to add values to it.
-    int Init(const EmPersistentValueList& values, bool removeUnusedValues);
+    int begin(EmPersistentValueList& values, bool removeUnusedValues);
+    int begin(EmPersistentValueBase* values[], size_t count, bool removeUnusedValues);
+    int begin(EmPSIterator& pIt, bool removeUnusedValues);
 
     // Checks if persistent state has been initialized (i.e. 'Init' call!)
-    bool IsInitialized() const {
-        return _isInitialized(false);
+    bool isInitialized() const {
+        return isInitialized_(false);
     }
 
     // Load the current persistent values into a list
     // Return the number of loaded values or -1 if persistent state has not been initialized.
     // NOTE:
     //  This method is dynamically allocating heap memory.
-    int Load(EmPersistentValueList& values);
-
-    // Iterate the persistent values one by one without generating a full elements list
-    // NOTE:
-    //  This method is dynamically allocating and deallocating heap memory.
-    bool Iterate(EmPersistentValueIterator& iterator);
+    int load(EmPersistentValueList& values);
 
     // Add a value to storage. 
     // This method will check if 'value' is already stored and set its 
     // current value taken from persistent state.
-    bool Add(EmPersistentValueBase& value);
+    bool add(EmPersistentValueBase& value);
 
     // Find this 'value' identified by its Id and Size.
     // If found the persistent state stored value is set to 'value'.
     // Return true if value has been fond in PS.
-    bool Find(EmPersistentValueBase& value);
+    bool find(EmPersistentValueBase& value);
 
     // Count the persistent state stored values or -1 if persistent state 
     // has not been initialized.
     // NOTE:
     //  This method is iterating trough all persistent state stored values.
-    int Count();
+    int count();
 
     // Clear the PS by resetting all its stored values
-    bool Clear();
+    bool clear();
 
 protected:   
-
     // Checks if persistent state has been initialized
-    bool _isInitialized(bool logError) const;
+    bool isInitialized_(bool logError) const;
 
     // Append a new value to storage
-    bool _appendValue(EmPersistentValueBase* pValue);
+    bool appendValue_(EmPersistentValueBase* pValue);
 
     // Performs a check if requested 'index' and 'size' are withint the PS boundaries
-    bool _indexCheck(ps_address_t index, ps_size_t size) const;
+    bool indexCheck_(ps_address_t index, ps_size_t size) const;
 
     // Read a byte from EEPROM
-    uint8_t _readByte(ps_address_t index) const;
+    uint8_t readByte_(ps_address_t index) const;
 
     // Read bytes from EEPROM
-    bool _readBytes(ps_address_t index, uint8_t* bytes, ps_size_t size) const;
+    bool readBytes_(ps_address_t index, uint8_t* bytes, ps_size_t size) const;
 
     // Update a byte to EEPROM
-    bool _updateByte(ps_address_t index, uint8_t byte) const;
+    bool updateByte_(ps_address_t index, uint8_t byte) const;
 
     // Update bytes to EEPROM
-    bool _updateBytes(ps_address_t index, 
+    bool updateBytes_(ps_address_t index, 
                       const uint8_t* bytes, 
                       ps_size_t size) const;
 
     // Find the matching id & size
-    bool _findMatch(ps_address_t& index, 
+    bool findMatch_(ps_address_t& index, 
                     const EmPersistentId& id, 
                     ps_size_t size) const;
                                    
     // Read the next PS id and size
-    bool _readNext(ps_address_t& index, 
+    bool readNext_(ps_address_t& index, 
                    EmPersistentId& id,
                    ps_size_t& size) const;
 
     // Create a new persistent value reading the next PS item
-    EmPersistentValueBase* _createNext(ps_address_t& index) const;
+    EmPersistentValueBase* createNext_(ps_address_t& index) const;
 
     // The first persistent value address
-    ps_address_t _firstPvAddress() const;
-    
+    ps_address_t firstPvAddress_() const;
+
+    // Get the next persistent value address
+    ps_address_t nextPvAddress_(ps_address_t index) const;
+
 private:
-    ps_address_t m_BeginIndex;
-    ps_address_t m_EndIndex;
-    ps_address_t m_NextPvAddress;
+    ps_address_t m_beginIndex;
+    ps_address_t m_endIndex;
+    ps_address_t m_nextPvAddress;
 };
+
+/***
+    The persistent value iterator used to iterate PS values
+***/
+class EmPersistentValueIterator: public EmPSIterator {
+public:
+    EmPersistentValueIterator(EmPersistentState& ps)
+      : m_ps(ps), m_pCurrentPv(nullptr) {}
+
+
+    virtual ~EmPersistentValueIterator() {
+        reset();
+    }
+
+    // No copy constructor or assignment operator
+    EmPersistentValueIterator(const EmPersistentValueIterator&) = delete;
+    EmPersistentValueIterator& operator=(const EmPersistentValueIterator&) = delete;
+
+    virtual void reset() override;
+    virtual bool next(EmPersistentValueBase*& pItem) override;
+
+protected:
+    EmPersistentState& m_ps;
+    EmPersistentValueBase* m_pCurrentPv;
+};    
 
 /***
     A unique ID assigned to a persistent value.
@@ -203,6 +236,9 @@ public:
     EmPersistentId(const char* id);
     EmPersistentId(const EmPersistentId& id);
 
+    // No assignment operator
+    EmPersistentId& operator=(const EmPersistentId&) = delete;
+
     // NOTE: keep destructor and class without virtual functions to avoid extra RAM consumption
     ~EmPersistentId() {
     }
@@ -210,7 +246,7 @@ public:
     char operator[](const int index ) const;
     
     bool operator==(const EmPersistentId& id) const { 
-        return 0 == memcmp(m_Id, id.m_Id, sizeof(m_Id)); 
+        return 0 == memcmp(m_id, id.m_id, sizeof(m_id)); 
     }
 
     bool operator!=(const EmPersistentId& id) const { 
@@ -218,26 +254,26 @@ public:
     }
 
     operator const char*() const { 
-        return GetId();
+        return getId();
     }
 
-    const char* GetId() const {
-        return m_Id;
+    const char* getId() const {
+        return m_id;
     }
     
 protected:
     EmPersistentId() {
-        memset(m_Id, 0, sizeof(m_Id));
+        memset(m_id, 0, sizeof(m_id));
     }
 
     // Read this object from PS
-    bool _read(const EmPersistentState& ps, ps_address_t index);
+    bool read_(const EmPersistentState& ps, ps_address_t index);
 
     // Store this object to PS
-    bool _store(const EmPersistentState& ps, ps_address_t index) const;
+    bool store_(const EmPersistentState& ps, ps_address_t index) const;
 
 private:
-    char m_Id[c_MaxLen+1];
+    char m_id[c_MaxLen+1];
 };
 
 /***
@@ -248,35 +284,39 @@ class EmPersistentValueBase {
     friend class EmPersistentValueIterator;
 public:    
     virtual ~EmPersistentValueBase() {
-        if (NULL != m_pValue) {
+        if (m_pValue != nullptr) {
             free(m_pValue);
         }
     }
 
-    const EmPersistentId& Id() const {
-        return m_Id;
+    // No copy constructor or assignment operator
+    EmPersistentValueBase(const EmPersistentValueBase&) = delete;
+    EmPersistentValueBase& operator=(const EmPersistentValueBase&) = delete;
+
+    const EmPersistentId& id() const {
+        return m_id;
     }
 
-    ps_address_t Address() const {
-        return m_Address;
+    ps_address_t address() const {
+        return m_address;
     }
 
-    ps_size_t Size() const {
-        return m_BufferSize;
+    ps_size_t size() const {
+        return m_bufferSize;
     }
 
-    bool IsStored() const {
-        return 0 != m_Address;
+    bool isStored() const {
+        return 0 != m_address;
     }
 
     // Checks if this two persistent values matches (i.e. have same Id and same Size)
-    bool Match(const EmPersistentValueBase& pv) const { 
-        return _match(Id(), pv.Id(), Size(), pv.Size());
+    bool match(const EmPersistentValueBase& pv) const { 
+        return match_(id(), pv.id(), size(), pv.size());
     }
 
 protected:
 
-    static bool _match(const EmPersistentId& id1,
+    static bool match_(const EmPersistentId& id1,
                        const EmPersistentId& id2,
                        ps_size_t size1,
                        ps_size_t size2) { 
@@ -287,62 +327,66 @@ protected:
                           const char* id,
                           ps_address_t address,
                           ps_size_t bufferSize,
-                          void* pInitValue = NULL);
+                          void* pInitValue = nullptr);
 
-    ps_address_t _idAddress() const {
-        return m_Address;
+    ps_address_t idAddress_() const {
+        return m_address;
     }
 
-    ps_address_t _sizeAddress() const {
-        return (ps_address_t)(m_Address+EmPersistentId::c_MaxLen);
+    ps_address_t sizeAddress_() const {
+        return (ps_address_t)(m_address+EmPersistentId::c_MaxLen);
     }
 
-    ps_address_t _valueAddress() const {
-        return (ps_address_t)(m_Address+EmPersistentId::c_MaxLen+
+    ps_address_t valueAddress_() const {
+        return (ps_address_t)(m_address+EmPersistentId::c_MaxLen+
                              (ps_address_t)sizeof(ps_size_t));
     }
 
-    ps_address_t _nextPvAddress() const {
-        return (ps_address_t)(m_Address+EmPersistentId::c_MaxLen+
-                              (ps_address_t)sizeof(ps_size_t)+m_BufferSize);
+    ps_address_t nextPvAddress_() const {
+        return (ps_address_t)(m_address+EmPersistentId::c_MaxLen+
+                              (ps_address_t)sizeof(ps_size_t)+m_bufferSize);
     }
 
     // Update the value to PS
-    bool _updateValue() const {
-        return m_Ps._updateBytes(_valueAddress(), (const uint8_t*)m_pValue, m_BufferSize);
+    bool updateValue_() const {
+        // Is this value stored (i.e. has been added to PS)?
+        if (!isStored()) {
+            return false;
+        }
+        return m_ps.updateBytes_(valueAddress_(), (const uint8_t*)m_pValue, m_bufferSize);
     }
 
-    virtual EmGetValueResult _getMem(void* pValue) const {
-        EmGetValueResult res = 0 == memcmp(pValue, m_pValue, m_BufferSize) ?
+    virtual EmGetValueResult getMem_(void* pValue) const {
+        EmGetValueResult res = 0 == memcmp(pValue, m_pValue, m_bufferSize) ?
                                EmGetValueResult::succeedEqualValue :
                                EmGetValueResult::succeedNotEqualValue;
-        memcpy(pValue, m_pValue, m_BufferSize);
+        memcpy(pValue, m_pValue, m_bufferSize);
         return res;
     }
 
-    virtual void _setMem(const void* pValue) {
-        memcpy(m_pValue, pValue, m_BufferSize);
+    virtual void setMem_(const void* pValue) {
+        memcpy(m_pValue, pValue, m_bufferSize);
     }
 
-    virtual bool _store() const;
+    virtual bool store_() const;
 
-    virtual void _copyFrom(EmPersistentValueBase* pPv) {
-        memcpy(m_pValue, pPv->m_pValue, m_BufferSize);
-        m_Address = pPv->m_Address;
+    virtual void copyFrom_(EmPersistentValueBase* pPv) {
+        memcpy(m_pValue, pPv->m_pValue, m_bufferSize);
+        m_address = pPv->m_address;
     }
 
 protected:
-    const EmPersistentState& m_Ps;
-    EmPersistentId m_Id;
-    ps_address_t m_Address;
-    ps_size_t m_BufferSize;
+    const EmPersistentState& m_ps;
+    EmPersistentId m_id;
+    ps_address_t m_address;
+    ps_size_t m_bufferSize;
     void* m_pValue;
 };
 
-inline bool _itemsMatch(const EmPersistentValueBase& pv1, 
+inline bool itemsMatch_(const EmPersistentValueBase& pv1, 
                         const EmPersistentValueBase& pv2) { 
     // This method is used to see if two items are the same in terms of EEPROM storage
-    return pv1.Match(pv2); 
+    return pv1.match(pv2); 
 }
 
 /***
@@ -361,30 +405,36 @@ public:
                             sizeof(T)) {
         // NOTE: 
         //  set memory directly instead calling 'SetValue' since Address is not set!
-        memcpy(m_pValue, &initValue, m_BufferSize);
+        memcpy(m_pValue, &initValue, m_bufferSize);
     }
 
-    virtual EmGetValueResult GetValue(T& value) const {
-        return _getMem(&value);
-    }
-        
-    virtual T GetValue() const {
+    virtual ~EmPersistentValue() = default;
+
+    // No copy constructor or assignment operator
+    EmPersistentValue(const EmPersistentValue&) = delete;
+    EmPersistentValue& operator=(const EmPersistentValue&) = delete;
+
+    virtual T getValue() const {
         T v = 0;
-        GetValue(v);
+        getValue(v);
         return v; 
     }
 
-    virtual bool SetValue(const T value) {
+    virtual EmGetValueResult getValue(T& value) const override {
+        return getMem_(&value);
+    }
+        
+    virtual bool setValue(const T value) override {
         // Avoid writing same value to EEPROM (only time consuming!)
-        if (Equals(value)) {
+        if (equals(value)) {
             return true;
         }
-        _setMem(&value);
-        return _updateValue();
+        setMem_(&value);
+        return updateValue_();
     }
 
-    virtual bool Equals(const T value) {
-        return 0 == memcmp(m_pValue, &value, m_BufferSize);
+    virtual bool equals(const T value) const {
+        return 0 == memcmp(m_pValue, &value, m_bufferSize);
     }
 
     virtual bool operator==(const T& other) const { 
@@ -396,7 +446,7 @@ public:
     }
 
     virtual operator T() const { 
-        return GetValue();
+        return getValue();
     }
 
     virtual operator void*() const { 
@@ -404,7 +454,7 @@ public:
     }
 
     virtual T operator =(T value) { 
-        SetValue(value);
+        setValue(value);
         return value; 
     }
 
@@ -429,33 +479,41 @@ typedef EmPersistentValue<uint64_t> EmPersistentUInt64;
 typedef EmPersistentValue<float> EmPersistentFloat;
 typedef EmPersistentValue<double> EmPersistentDouble;
 
-class EmPersistentString: public EmPersistentValue<char*> {
+class EmPersistentString: public EmPersistentValueBase, public EmValue<char*> {
+    friend class EmPersistentState;
 public:
     EmPersistentString(const EmPersistentState& ps,
                        const char* id,
                        ps_size_t maxTextLen,
                        const char* initValue)
-    : EmPersistentValue(ps, id, (ps_address_t)0, (ps_size_t)(maxTextLen+1), NULL) {
+    : EmPersistentValueBase(ps, id, 0, maxTextLen+1) {
         // NOTE:
         //   We NEED to copy initValue within this constructor and NOT base one!
-        memcpy(m_pValue, initValue, _valueSize(initValue));
+        memcpy(m_pValue, initValue, valueSize_(initValue));
     }
-    
-    virtual EmGetValueResult GetValue(char* value) const {
+
+    virtual const char* getValue() const {
+        return (const char*)m_pValue;
+    }
+
+    virtual EmGetValueResult getValue(char* value) const {
         return _getMem((void*)value);
     }
 
-    virtual bool SetValue(const char* value) {
+    virtual bool setValue(const char* value) {
         // Avoid writing same value to EEPROM (only time consuming!)
-        if (Equals(value)) {
+        if (equals(value)) {
             return true;
         }
-        _setMem(value);
-        return _updateValue();
+        setMem_(value);
+        return updateValue_();
     }
 
-    virtual bool Equals(const char* value) {
-        return 0 == memcmp(m_pValue, value, _valueSize(value));
+    virtual bool equals(const char* value) const {
+        if (value == nullptr) {
+            return 0 == ((const char*)m_pValue)[0];
+        }
+        return 0 == strcmp((const char*)m_pValue, value);
     }
 
     virtual operator const char*() const { 
@@ -463,77 +521,40 @@ public:
     }
 
     virtual char* operator =(const char* value) {         
-        SetValue(value);
+        setValue(value);
         return (char*)m_pValue;
     }
 
+    virtual bool operator==(const char* other) const { 
+        return this->equals(other); 
+    }
+
+    virtual bool operator!=(const char* other) const { 
+        return !this->equals(other); 
+    }    
+
 protected:
     virtual EmGetValueResult _getMem(void* pValue) const {
-        EmGetValueResult res = 0 == memcmp(pValue, m_pValue, _valueSize((const char*)pValue)) ?
+        EmGetValueResult res = 0 == memcmp(pValue, m_pValue, valueSize_((const char*)pValue)) ?
                                EmGetValueResult::succeedEqualValue : 
                                EmGetValueResult::succeedNotEqualValue;
-        memcpy(pValue, m_pValue, _valueSize((const char*)pValue));
+        memcpy(pValue, m_pValue, valueSize_((const char*)pValue));
         return res;
     }
 
-    virtual void _setMem(const void* pValue) {
-        memcpy(m_pValue, pValue, _valueSize((const char*)pValue));
+    virtual void setMem_(const void* pValue) {
+        memcpy(m_pValue, pValue, valueSize_((const char*)pValue));
     }
 
-    virtual ps_size_t _valueSize(const char* pValue) const {
-        if (NULL == pValue) {
+    virtual ps_size_t valueSize_(const char* pValue) const {
+        if (nullptr == pValue) {
             return 0;
         }
         // +1 -> Need to set the string terminator as well
         ps_size_t valueSize = (ps_size_t)(strlen(pValue)+1); 
         // -1 -> Need to leave the string terminator (i.e. max length reached!)
-        return (ps_size_t)MIN(m_BufferSize-1, valueSize);
+        return (ps_size_t)MIN(m_bufferSize-1, valueSize);
     }
 };
 
-/***
-    The persistent values iterator
-***/
-class EmPersistentValueIterator  {
-    friend class EmPersistentState;
-public:
-    EmPersistentValueIterator() 
-     : m_pItem(NULL), 
-       m_EndReached(false) {}
-
-    ~EmPersistentValueIterator() {
-        Reset();
-    }
-
-    void Reset() {
-        if (NULL != m_pItem) {
-            delete m_pItem;
-        }
-        m_pItem = NULL;
-        m_EndReached = false;
-    }
-
-    operator EmPersistentValueBase*() {
-        return m_pItem;
-    }
-
-    EmPersistentValueBase* Item() {
-        return m_pItem;
-    }
-
-    bool EndReached() {
-        return m_EndReached;
-    }
-
-protected:
-
-    void _setItem(EmPersistentValueBase* pItem) {
-        Reset();
-        m_pItem = pItem;        
-        m_EndReached = NULL == pItem;
-    }
-
-private:
-    EmPersistentValueBase* m_pItem;
-    bool m_EndReached;
-};
+#endif