+++ CHANGELOG.md
@@ -3,5 +3,8 @@
 # 1.0.3
 - first releases with several adjustments and fixes
 
-# 1.0.4
-- Added explicit 'GetValue()' in 'EmPersistentValue' class 
+# 2.0.0
+- Aligned to new EmCore 2.0.0 code restyling
+- Added explicit 'getValue()' in 'EmPersistentValue' class 
+- Added EmIterator support (i.e. initialization with both list or array)
+- Added support for ESP32 EEPROM implementation (NOTE: prefer using EmCore's EmStorage)